requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
openpyxl
python-dateutil